requirements.txt
python-dotenv==1.0.1
openai==1.59.7
psycopg[binary]==3.3.2
streamlit==1.41.1
reportlab
streamlit-lottie
requests
